Restructure modules into subpackages
