+++ src/cogniweave/llms/chat.py
@@ -26,7 +26,7 @@ from langchain_core.tools import BaseTool
 from langchain_openai import ChatOpenAI as BaseChatOpenAI
 from pydantic import BaseModel, Field, model_validator
 
-from cogniweave.prompt_values import MultilingualSystemPromptValue
+from cogniweave.prompt_values.base import MultilingualSystemPromptValue
 from cogniweave.typing import (
     MessageLikeRepresentation,
     Output,
