+++ src/cogniweave/prompts/__init__.py
@@ -1,0 +1,13 @@
+from .message_templates import (
+    MessageSegmentsPlaceholder,
+    RichHumanMessagePromptTemplate,
+    RichAIMessagePromptTemplate,
+    RichSystemMessagePromptTemplate,
+)
+
+__all__ = [
+    "MessageSegmentsPlaceholder",
+    "RichHumanMessagePromptTemplate",
+    "RichAIMessagePromptTemplate",
+    "RichSystemMessagePromptTemplate",
+]
