+++ src/cogniweave/time_splitter/__init__.py
@@ -1,0 +1,3 @@
+from .base import BaseTimeSplitter
+
+__all__ = ["BaseTimeSplitter"]
